+++ src/Filter/Windows1252ToUtf8.php
@@ -5,34 +5,26 @@ declare(strict_types=1);
 namespace MarcusJaschen\Collmex\Filter;
 
 use ForceUTF8\Encoding;
+use InvalidArgumentException;
 
 /**
  * Filter to convert Windows 1252 to UTF-8 encoding.
  *
  * @author   Marcus Jaschen <mail@marcusjaschen.de>
  */
-class Windows1252ToUtf8 implements FilterInterface
+class Windows1252ToUtf8 extends AbstractFilter
 {
     /**
-     * @param string $text
-     *
-     * @return string
+     * @throws InvalidArgumentException
      */
     public function filterString(string $text): string
     {
-        return Encoding::toUTF8($text);
-    }
+        $result = mb_convert_encoding($text, 'UTF-8', 'Windows-1252');
 
-    /**
-     * @param string[] $input
-     *
-     * @return string[]
-     */
-    public function filterArray(array $input): array
-    {
-        // The Encoding methods are array-aware so we can just drop our input
-        // into the conversion method.
+        if ($result === false) {
+            throw new InvalidArgumentException('Cannot convert string to UTF-8', 8609163127);
+        }
 
-        return (array)Encoding::toUTF8($input);
+        return $result;
     }
 }