+++ src/Type/CustomerOrder.php
@@ -17,9 +17,9 @@ namespace MarcusJaschen\Collmex\Type;
  * @property $customer_id
  * @property $customer_salutation
  * @property $customer_title
- * @property $customer_forename
+ * @property $customer_firstname
  * @property $customer_lastname
- * @property $customer_firm
+ * @property $customer_company
  * @property $customer_department
  * @property $customer_street
  * @property $customer_zipcode
@@ -68,9 +68,9 @@ namespace MarcusJaschen\Collmex\Type;
  * @property $delivery_conditions_additional
  * @property $delivery_salutation
  * @property $delivery_title
- * @property $delivery_forename
+ * @property $delivery_firstname
  * @property $delivery_lastname
- * @property $delivery_firm
+ * @property $delivery_company
  * @property $delivery_department
  * @property $delivery_street
  * @property $delivery_zipcode
@@ -238,10 +238,10 @@ class CustomerOrder extends AbstractType implements TypeInterface
         'customer_id' => null,
         'customer_salutation' => null,
         'customer_title' => null,
-        'customer_forename' => null,
+        'customer_firstname' => null,
         // 10
         'customer_lastname' => null,
-        'customer_firm' => null,
+        'customer_company' => null,
         'customer_department' => null,
         'customer_street' => null,
         'customer_zipcode' => null,
@@ -295,9 +295,9 @@ class CustomerOrder extends AbstractType implements TypeInterface
         'delivery_salutation' => null,
         'delivery_title' => null,
         // 60
-        'delivery_forename' => null,
+        'delivery_firstname' => null,
         'delivery_lastname' => null,
-        'delivery_firm' => null,
+        'delivery_company' => null,
         'delivery_department' => null,
         'delivery_street' => null,
         'delivery_zipcode' => null,