+++ src/Type/Invoice.php
@@ -18,9 +18,9 @@ namespace MarcusJaschen\Collmex\Type;
  * @property $customer_id
  * @property $customer_salutation
  * @property $customer_title
- * @property $customer_forename
+ * @property $customer_firstname
  * @property $customer_lastname
- * @property $customer_firm
+ * @property $customer_company
  * @property $customer_department
  * @property $customer_street
  * @property $customer_zipcode
@@ -65,9 +65,9 @@ namespace MarcusJaschen\Collmex\Type;
  * @property $delivery_conditions_additional
  * @property $delivery_salutation
  * @property $delivery_title
- * @property $delivery_forename
+ * @property $delivery_firstname
  * @property $delivery_lastname
- * @property $delivery_firm
+ * @property $delivery_company
  * @property $delivery_department
  * @property $delivery_street
  * @property $delivery_zipcode
@@ -219,9 +219,9 @@ class Invoice extends AbstractType implements TypeInterface
         'customer_salutation' => null,
         'customer_title' => null,
         // 10
-        'customer_forename' => null,
+        'customer_firstname' => null,
         'customer_lastname' => null,
-        'customer_firm' => null,
+        'customer_company' => null,
         'customer_department' => null,
         'customer_street' => null,
         // 15
@@ -275,9 +275,9 @@ class Invoice extends AbstractType implements TypeInterface
         // 55
         'delivery_salutation' => null,
         'delivery_title' => null,
-        'delivery_forename' => null,
+        'delivery_firstname' => null,
         'delivery_lastname' => null,
-        'delivery_firm' => null,
+        'delivery_company' => null,
         // 60
         'delivery_department' => null,
         'delivery_street' => null,